quiz-video-backend/pyproject.toml
[project]
name = "quiz-video-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "numpy>=1.26",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=2.7.0",
  "faster-whisper>=1.0.3",
  "aiofiles>=23.2.1",
]

[tool.uvicorn]
factory = false
